pyproject.toml
[project]
name = "loan-approval-bot"
version = "0.1.0"
description = "Loan Approval Recommendation AI agent using LangGraph"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2,<2.0",
    "langchain-core>=1.3,<2.0",
    "langchain-openai>=1.2,<2.0",
    "langgraph>=0.2,<2.0",
    "pydantic>=2.11,<3.0",
    "python-dotenv>=1.0,<2.0",
    "gradio>=5.0,<7.0",
    "pypdf>=6.10.2",
    "fastapi>=0.100,<1.0",
    "uvicorn>=0.23,<1.0",
]
